requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
matplotlib>=3.8.0
PyYAML>=6.0
pytest>=8.0.0
requests>=2.31
